requirements.txt
pandas
numpy
vectorbt
dateparser 
yfinance
kaleido
nbformat>=4.2.0
ta
supabase
python-dotenv
tqdm
pandas-datareader
setuptools